+++ central/certs/handlers/http_handler.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/stackrox/rox/central/certs/export"
 	"github.com/stackrox/rox/central/systeminfo/listener"
+	"github.com/stackrox/rox/pkg/httputil"
 	"github.com/stackrox/rox/pkg/logging"
 )
 
@@ -23,12 +24,6 @@ func BackupCerts(backupListener listener.BackupListener) http.Handler {
 	return dumpCerts(backupListener)
 }
 
-func logAndWriteErrorMsg(w http.ResponseWriter, code int, t string, args ...interface{}) {
-	errMsg := fmt.Sprintf(t, args...)
-	log.Error(errMsg)
-	http.Error(w, errMsg, code)
-}
-
 func dumpCerts(backupListener listener.BackupListener) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
 		log.Info("Starting cert backup ...")
@@ -38,7 +33,7 @@ func dumpCerts(backupListener listener.BackupListener) http.HandlerFunc {
 		w.Header().Set("Content-Disposition", fmt.Sprintf("attachment; filename=\"%s\"", filename))
 
 		if err := export.BackupCerts(req.Context(), backupListener, w); err != nil {
-			logAndWriteErrorMsg(w, http.StatusInternalServerError, "could not create cert backup: %v", err)
+			httputil.WriteErrorf(w, http.StatusInternalServerError, "could not create cert backup: %v", err)
 			return
 		}
 		log.Info("Cert backup completed")